Remove setup.rst
There's no such GillesPy2 module, so there's nothing to document in
the code itself.  If this was meant to document ../setup.py, then I'm
not sure it's worth it -- it would be unusual compared to other
sphinx-based Python docs I've seen.  I think most probably setup.rst
was the result of running a sphinx auto generator of some kind and did
not reflect a real module that had to be documented.