Conditionally load CythonSSAsolver
This code change makes it so that if the user does not have Cython or
pyximport installed, the code won't fail with a load error.  This also
makes it possible for `run_tests.py` to load all the `test_XYZ`
functions without fear.  (Previously, `test_cython_ssa_solver` was
commented out, presumably because it caused a load failure if Cython
wasn't available.)