Fix inconsistent propensity formula orderings
When I ran the unit tests, there was a failure in
`test_ode_propensity` in `test/test_mode.py`.  The error was:

```
FAIL: test_ode_propensity (test_model.TestModel)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/Users/mhucka/project-files/stochss/repos/GillesPy2/test/test_model.py", line 151, in test_ode_propensity
    self.assertEqual(model.listOfReactions['r3'].ode_propensity_function, 'rate*A*B')
AssertionError: 'rate*B*A' != 'rate*A*B'
- rate*B*A
?       --
+ rate*A*B
?      ++
```

The problem was due to the way that propensity formulas were sorted in
`Reaction.sanitized_propensity_function(...)`.  The method sorted by the
length of the name of the variables in the formula, but this will not yield a
consistent ordering across different platforms when the names are of the same
length.

My solution was to implement proper Python object ordering for `Species`,
`Parameter` and a couple of other object classes.  The implementation uses a
simple base class, `SortableObject`, that defines Python methods for things
like `__eq__(...)`, `__le__(...)`, etc.  The class definitions for `Model`,
`Species` and other things then use that base class, which makes those object
instances comparable using normal Python functions like `sorted(...)`.  From
there, it was simple to change `Reaction.sanitized_propensity_function(...)`
to sort the variables in the propensity formula in a consistent way based on
alphabetic sort.

For good measure, I looked for other places in the code that used `for` loops
to iterate over objects and append things to lists, and wrapped the lists
with `sorted(...)`.  I don't know for sure whether those were actually
necessary, but it seemed safer, to ensure that they were more likely to lead
to consistent behavior.