Improve wrappers for SBML Test Runner
Several changes are incorporated here.

First, this file is a polyglot script valid in both sh and Python.
This approach is used to solve the problem that we don't know whether
the user's "python" is Python 2 or 3, and thus we can't use an
explicit path to python3 in the #! line, nor can we use #!/usr/bin/env
python3.  This polyglot script operates by running in sh or bash at
first, and then it execs Python on the script.  The rest of the file
is Python.

Second, the script adds help text via a new -h (or --help)
command-line option.

Third, there's a tiny bit of tweaking in how command-line arguments
are handled.