Rewrite ODE solver for boundaryCondition & constant Species
The primary behavior change is that this now supports the SBML
attributes of "boundaryCondition" and "constant" on species.

In order to be able to figure out how to implement support for this, I
ended up rewriting much of the code in order to (I hope) increase
clarity, remove some (seemingly) inefficient or convoluted parts, and
add a lot of comments.