Refactor to support boundary condition & streamline code
The changes for species boundary conditions and constant species were
small and pretty simple, but to figure out what had to be done, I had
to understand all the code first.  Once I understood most of it, I
realized some of it could be written to be simpler, which helped my
understanding.  And here we are.