Fixed (FOR REAL) the translation bug where object variable names would also be translated.
The issue was not the implementation, but
instead a bug within the recursive translate function. When recursing
into an object, it would recurse on the backing __dict__, translating
both keys and values. Instead, iterate through each variable in the
backing __dict__ and translate the values, not the keys.

Due to this change, variable names on objects which inherit from Jsonify
are protected from change.